Cli support exporting document
