+++ core/src/index.ts
@@ -10,3 +10,4 @@ export * as markupConvertor from "./markupConvertor";
 export * as variableReplacer from "./variableReplacer";
 export * as outputExtractor from "./outputExtractor";
 export * as testHelper from "./testHelper";
+export * as docHtml from "./docHtml";
